+++ src/main/java/recsys/evaluator/WeightedRMSRecommenderCrossEvaluator.java
@@ -9,10 +9,10 @@ import recsys.evaluator.abstr.AbstractDifferenceRecommenderCrossEvaluator;
 public final class WeightedRMSRecommenderCrossEvaluator extends AbstractDifferenceRecommenderCrossEvaluator {
 
 	private RunningAverage average;
-	private float numberOfDistinctPreferenceValues;
-	private float preferenceValueCorrection;
+	private Double numberOfDistinctPreferenceValues;
+	private Double preferenceValueCorrection;
 
-	public WeightedRMSRecommenderCrossEvaluator(float numberOfDistinctPreferenceValues, float preferenceValueCorrection) {
+	public WeightedRMSRecommenderCrossEvaluator(Double numberOfDistinctPreferenceValues, Double preferenceValueCorrection) {
 		this.numberOfDistinctPreferenceValues = numberOfDistinctPreferenceValues;
 		this.preferenceValueCorrection = preferenceValueCorrection;
 	}
@@ -23,16 +23,16 @@ public final class WeightedRMSRecommenderCrossEvaluator extends AbstractDifferen
 	}
 
 	@Override
-	protected void processOneEstimate(float estimatedPreference, Preference realPref) {
-		float realPrefValue = realPref.getValue();
-		float weight = (realPrefValue + preferenceValueCorrection) / numberOfDistinctPreferenceValues;
-		double diff = weight * (realPrefValue - estimatedPreference);
+	protected void processOneEstimate(Double estimatedPreference, Preference realPref) {
+		Double realPrefValue = realPref.getValue();
+		Double weight = (realPrefValue + preferenceValueCorrection) / numberOfDistinctPreferenceValues;
+		Double diff = weight * (realPrefValue - estimatedPreference);
 		average.addDatum(diff * diff);
 	}
 
 	@Override
-	protected double computeFinalEvaluation() {
-		return Math.sqrt(average.getAverage());
+	protected Double computeFinalEvaluation() {
+		return (double) Math.sqrt(average.getAverage());
 	}
 
 	@Override
