+++ src/main/java/recsys/notebooks/recommender/NotebooksLocalClassificationRecommender.java
@@ -14,11 +14,11 @@ public abstract class NotebooksLocalClassificationRecommender extends NotebooksC
 	}
 
 	@Override
-	public float estimatePreference(long userID, long itemID) throws TasteException {
+	public Double estimatePreference(Integer userID, Integer itemID) throws TasteException {
 		try {
-			double localResult = getLocalResult(userID, itemID);
+			Double localResult = getLocalResult(userID, itemID);
 			
-			return (float) localResult;
+			return (double) localResult;
 		} catch (Exception e) {
 			throw new TasteException(e);
 		}
